+++ src/prompt.rs
@@ -1,6 +1,6 @@
 use super::database::*;
 use super::model::*;
-use dialoguer::{Confirm, MultiSelect, Select};
+use dialoguer::{Confirm, Editor, FuzzySelect, MultiSelect};
 use simple_error::SimpleResult;
 use sqlx::sqlite::SqliteConnection;
 use std::path::PathBuf;
@@ -40,6 +40,29 @@ pub async fn prompt_for_systems(
         .filter(|(i, _)| indices.contains(i))
         .map(|(_, system)| system)
         .collect())
+}
+
+pub async fn prompt_for_system(
+    connection: &mut SqliteConnection,
+    default: Option<usize>,
+) -> SimpleResult<System> {
+    let mut systems = find_systems(connection).await;
+    match systems.len() {
+        0 => bail!("No available system"),
+        1 => Ok(systems.remove(0)),
+        _ => {
+            let index = select(
+                &systems
+                    .iter()
+                    .map(|system| &system.name)
+                    .collect::<Vec<&String>>(),
+                "Please select a system",
+                default,
+                None,
+            )?;
+            Ok(systems.remove(index))
+        }
+    }
 }
 
 pub async fn prompt_for_system_like(
@@ -88,23 +111,29 @@ pub fn prompt_for_games(games: Vec<Game>, all: bool) -> SimpleResult<Vec<Game>> 
         .collect())
 }
 
-pub fn prompt_for_game(games: &[Game]) -> SimpleResult<Option<&Game>> {
-    let index = select_opt(
-        &games
-            .iter()
-            .map(|game| &game.name)
-            .collect::<Vec<&String>>(),
-        "Please select a game",
-        None,
-        Some(10),
-    )?;
-    Ok(index.map(|i| games.get(i).unwrap()))
+pub fn prompt_for_game(games: &[Game], default: Option<usize>) -> SimpleResult<Option<&Game>> {
+    match games.len() {
+        0 => bail!("No available rom"),
+        1 => Ok(games.get(0)),
+        _ => {
+            let index = select_opt(
+                &games
+                    .iter()
+                    .map(|game| &game.name)
+                    .collect::<Vec<&String>>(),
+                "Please select a game",
+                default,
+                Some(10),
+            )?;
+            Ok(index.map(|i| games.get(i).unwrap()))
+        }
+    }
 }
 
-pub fn prompt_for_rom(roms: &mut Vec<Rom>, default: Option<usize>) -> SimpleResult<Option<Rom>> {
+pub fn prompt_for_rom(roms: &[Rom], default: Option<usize>) -> SimpleResult<Option<&Rom>> {
     match roms.len() {
         0 => bail!("No available rom"),
-        1 => Ok(Some(roms.remove(0))),
+        1 => Ok(roms.get(0)),
         _ => {
             let index = select_opt(
                 &roms.iter().map(|rom| &rom.name).collect::<Vec<&String>>(),
@@ -112,7 +141,7 @@ pub fn prompt_for_rom(roms: &mut Vec<Rom>, default: Option<usize>) -> SimpleResu
                 default,
                 Some(10),
             )?;
-            Ok(index.map(|i| roms.remove(i)))
+            Ok(index.map(|i| roms.get(i).unwrap()))
         }
     }
 }
@@ -188,6 +217,10 @@ pub async fn prompt_for_parent_romfile(
     Ok(index.map(|index| romfiles.remove(index)))
 }
 
+pub fn prompt_for_name(prompt: &str) -> SimpleResult<Option<String>> {
+    Ok(editor(prompt)?)
+}
+
 pub fn confirm(default: bool) -> SimpleResult<bool> {
     Ok(try_with!(
         Confirm::new()
@@ -198,13 +231,20 @@ pub fn confirm(default: bool) -> SimpleResult<bool> {
     ))
 }
 
+pub fn editor(prompt: &str) -> SimpleResult<Option<String>> {
+    Ok(try_with!(
+        Editor::new().edit(prompt),
+        "Failed to get user input"
+    ))
+}
+
 pub fn select<T: ToString>(
     items: &[T],
     prompt: &str,
     default: Option<usize>,
     max_length: Option<usize>,
 ) -> SimpleResult<usize> {
-    let mut select = Select::new();
+    let mut select = FuzzySelect::new();
     select = select.items(items).with_prompt(prompt);
     if let Some(default) = default {
         select = select.default(default);
@@ -221,7 +261,7 @@ pub fn select_opt<T: ToString>(
     default: Option<usize>,
     max_length: Option<usize>,
 ) -> SimpleResult<Option<usize>> {
-    let mut select = Select::new();
+    let mut select = FuzzySelect::new();
     select = select.items(items).with_prompt(prompt);
     if let Some(default) = default {
         select = select.default(default);