+++ src/sort_roms/test_path_subfolder_alpha_other.rs
@@ -50,6 +50,7 @@ async fn test() {
         path: String::from("42.rom"),
         size: 0,
         parent_id: None,
+        romfile_type: RomfileType::Romfile as i64,
     };
 
     // when