+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add a `-x` flag to `import-roms` to extract top-level archives before importing their contents
 - Allow ordering CHD compression algorithms
 - Use the new `splitbin` feature in chdman 0.265 and higher, fixes Dreamcast CHD imports
+- Add a new `import-patches` subcommand
 
 # 0.19.0